+++ pkg/kubelet/network/dns/OWNERS
@@ -1,0 +1,4 @@
+approvers:
+- sig-network-approvers
+reviewers:
+- sig-network-reviewers
